fix(win-x64): T-win-50023 __negtf2 sign mask — 1UL<<63 is 32-bit on LLP64 arm64-Windows so it flipped bit 95 not the sign bit 127; use (uint64_t)1<<63 (byte-identical on LP64) [win-x64]

+++ runtime/lib/float128.c
@@ -347,7 +347,7 @@ f128_t __divtf3(f128_t fa, f128_t fb) {
 }
 
 f128_t __negtf2(f128_t f) {
-	((u128_t *)&f)->x1 ^= 1UL << 63;
+	((u128_t *)&f)->x1 ^= (uint64_t)1 << 63;
 	return f;
 }
